fix sqn type error

+++ TradingEnvironment.py
@@ -394,7 +394,7 @@ class TradingEnvironment:
         expected_value = Decimal(win_rate) * (1 + odds) - 1
         mdd = np.min(self.trading_profits)
         mar = np.sum(self.trading_profits) / mdd
-        sqn = np.mean(self.trading_profits) / np.std(self.trading_profits) * np.sqrt(len(self.trading_profits))
+        sqn = Decimal(np.mean(self.trading_profits) / np.std(self.trading_profits)) * Decimal(np.sqrt(len(self.trading_profits)))
 
         print(f"{self.ticker} trading record")
         print(self.initial_balance)
